fix: real Freedom avatars + NFT query crash fix
- Add AvatarImage component using gateway.freedom.world/api/freedom-wallet/profile/image/{fdvId}
- Fallback to ui-avatars.com on error
- Fix received_nft SUM crash (column is text, needs ::int cast)
- Update LeaderboardTable + player profile page to use real avatars

+++ lib/queries/player.ts
@@ -111,7 +111,7 @@ export async function getPlayerCombatStats(fdvId: number) {
 
 export async function getTotalNFTs(fdvId: number): Promise<number> {
   const rows = await neonSql`
-    SELECT COALESCE(SUM(received_nft), 0)::int as total_nfts
+    SELECT COALESCE(SUM(received_nft::int), 0)::int as total_nfts
     FROM web_card_game_prod.game_match_resulted WHERE player_fdv_id = ${fdvId}
   `;
   return (rows[0] as { total_nfts: number })?.total_nfts || 0;
